requirements.txt
streamlit==1.34.0
requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
google-generativeai==0.5.4
